Migrated all pages and component to typescri[t
