+++ frontend/src/layouts/NavBar.tsx
