+++ frontend/src/pages/Home.tsx
@@ -3,11 +3,16 @@ import LoginForm from "../components/LoginForm";
 import RegisterForm from "../components/RegisterForm";
 import TabComponent from "../components/TabComponent";
 
+export enum TabsEnum {
+  Login = "Login",
+  Register = "Register",
+}
+
 const Home = () => {
   const tabs = ["Login", "Register"];
 
-  const [activeIndex, setActiveIndex] = useState(tabs[0]);
-  const handleTabChange = (tab) => {
+  const [activeIndex, setActiveIndex] = useState<TabsEnum>(TabsEnum.Login);
+  const handleTabChange = (tab: TabsEnum) => {
     setActiveIndex(tab);
   };
   return (
@@ -24,7 +29,6 @@ const Home = () => {
         <div className="max-w-md">
           <h1 className="mb-5 text-5xl font-bold">Welcome there</h1>
           <p className="mb-5">
-            {" "}
             Discover a world of possibilities with our platform.
           </p>
